Yay for conflicts galore (wax removal, std:: purge, tileidx changes).

+++ crawl-ref/source/spl-clouds.cc
@@ -48,7 +48,7 @@ static void _burn_tree(coord_def pos)
 spret_type conjure_flame(int pow, const coord_def& where, bool fail)
 {
     // FIXME: This would be better handled by a flag to enforce max range.
-    if (distance(where, you.pos()) > dist_range(spell_range(SPELL_CONJURE_FLAME,
+    if (distance2(where, you.pos()) > dist_range(spell_range(SPELL_CONJURE_FLAME,
                                                       pow))
         || !in_bounds(where))
     {
@@ -110,13 +110,13 @@ spret_type conjure_flame(int pow, const coord_def& where, bool fail)
         // Reinforce the cloud - but not too much.
         // It must be a fire cloud from a previous test.
         mpr("The fire roars with new energy!");
-        const int extra_dur = 2 + std::min(random2(pow) / 2, 20);
+        const int extra_dur = 2 + min(random2(pow) / 2, 20);
         env.cloud[cloud].decay += extra_dur * 5;
         env.cloud[cloud].set_whose(KC_YOU);
     }
     else
     {
-        const int durat = std::min(5 + (random2(pow)/2) + (random2(pow)/2), 23);
+        const int durat = min(5 + (random2(pow)/2) + (random2(pow)/2), 23);
         place_cloud(CLOUD_FIRE, where, durat, &you);
         mpr("The fire roars!");
     }
@@ -169,7 +169,7 @@ spret_type stinking_cloud(int pow, bolt &beem, bool fail)
 spret_type cast_big_c(int pow, cloud_type cty, const actor *caster, bolt &beam,
                       bool fail)
 {
-    if (distance(beam.target, you.pos()) > dist_range(beam.range)
+    if (distance2(beam.target, you.pos()) > dist_range(beam.range)
         || !in_bounds(beam.target))
     {
         mpr("That is beyond the maximum range.");
@@ -191,7 +191,7 @@ spret_type cast_big_c(int pow, cloud_type cty, const actor *caster, bolt &beam,
 
 static int _make_a_normal_cloud(coord_def where, int pow, int spread_rate,
                                 cloud_type ctype, const actor *agent, int colour,
-                                std::string name, std::string tile, int excl_rad)
+                                string name, string tile, int excl_rad)
 {
     place_cloud(ctype, where,
                 (3 + random2(pow / 4) + random2(pow / 4) + random2(pow / 4)),
@@ -202,7 +202,7 @@ static int _make_a_normal_cloud(coord_def where, int pow, int spread_rate,
 
 void big_cloud(cloud_type cl_type, const actor *agent,
                const coord_def& where, int pow, int size, int spread_rate,
-               int colour, std::string name, std::string tile)
+               int colour, string name, string tile)
 {
     apply_area_cloud(_make_a_normal_cloud, where, pow, size,
                      cl_type, agent, spread_rate, colour, name, tile,
@@ -237,9 +237,11 @@ void manage_fire_shield(int delay)
         you.duration[DUR_FIRE_SHIELD] = 0;
 
     // Remove fire clouds on top of you
-    if (env.cgrid(you.pos()) != EMPTY_CLOUD)
-        if (env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
-            delete_cloud_at(you.pos());
+    if (env.cgrid(you.pos()) != EMPTY_CLOUD
+        && env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
+    {
+        delete_cloud_at(you.pos());
+    }
 
     if (!you.duration[DUR_FIRE_SHIELD])
     {
@@ -296,273 +298,6 @@ void corpse_rot(actor* caster)
         mpr("You smell decay.");
 
     // Should make zombies decay into skeletons?
-}
-
-// Returns a vector of cloud types created by this potion type.
-// FIXME: Heavily duplicated code.
-static std::vector<int> _get_evaporate_result(int potion)
-{
-    std::vector <int> beams;
-    bool random_potion = false;
-    switch (potion)
-    {
-    case POT_STRONG_POISON:
-    case POT_POISON:
-        beams.push_back(BEAM_POTION_POISON);
-        break;
-
-    case POT_DEGENERATION:
-        beams.push_back(BEAM_POTION_POISON);
-        beams.push_back(BEAM_POTION_MIASMA);
-        break;
-
-    case POT_DECAY:
-        beams.push_back(BEAM_POTION_MIASMA);
-        break;
-
-    case POT_PARALYSIS:
-    case POT_CONFUSION:
-    case POT_SLOWING:
-        beams.push_back(BEAM_POTION_MEPHITIC);
-        break;
-
-    case POT_WATER:
-    case POT_PORRIDGE:
-        beams.push_back(BEAM_POTION_STEAM);
-        break;
-
-    case POT_BLOOD:
-    case POT_BLOOD_COAGULATED:
-        beams.push_back(BEAM_POTION_MEPHITIC);
-        // deliberate fall through
-    case POT_BERSERK_RAGE:
-        beams.push_back(BEAM_POTION_FIRE);
-        beams.push_back(BEAM_POTION_STEAM);
-        break;
-
-    case POT_MUTATION:
-        beams.push_back(BEAM_POTION_MUTAGENIC);
-        // deliberate fall-through
-    case POT_GAIN_STRENGTH:
-    case POT_GAIN_DEXTERITY:
-    case POT_GAIN_INTELLIGENCE:
-    case POT_EXPERIENCE:
-    case POT_MAGIC:
-        beams.push_back(BEAM_POTION_FIRE);
-        beams.push_back(BEAM_POTION_COLD);
-        beams.push_back(BEAM_POTION_POISON);
-        beams.push_back(BEAM_POTION_MIASMA);
-        random_potion = true;
-        break;
-
-    default:
-        beams.push_back(BEAM_POTION_FIRE);
-        beams.push_back(BEAM_POTION_MEPHITIC);
-        beams.push_back(BEAM_POTION_COLD);
-        beams.push_back(BEAM_POTION_POISON);
-        beams.push_back(BEAM_POTION_BLUE_SMOKE);
-        beams.push_back(BEAM_POTION_STEAM);
-        random_potion = true;
-        break;
-    }
-
-    std::vector<int> clouds;
-    for (unsigned int k = 0; k < beams.size(); ++k)
-        clouds.push_back(beam2cloud((beam_type) beams[k]));
-
-    if (random_potion)
-    {
-        // handled in beam.cc
-        clouds.push_back(CLOUD_FIRE);
-        clouds.push_back(CLOUD_MEPHITIC);
-        clouds.push_back(CLOUD_COLD);
-        clouds.push_back(CLOUD_POISON);
-        clouds.push_back(CLOUD_BLUE_SMOKE);
-        clouds.push_back(CLOUD_STEAM);
-    }
-
-    return clouds;
-}
-
-// Returns a comma-separated list of all cloud types potentially created
-// by this potion type. Doesn't respect the different probabilities.
-std::string get_evaporate_result_list(int potion)
-{
-    std::vector<int> clouds = _get_evaporate_result(potion);
-    std::sort(clouds.begin(), clouds.end());
-
-    std::vector<std::string> clouds_list;
-
-    int old_cloud = -1;
-    for (unsigned int k = 0; k < clouds.size(); ++k)
-    {
-        const int new_cloud = clouds[k];
-        if (new_cloud == old_cloud)
-            continue;
-
-        // This relies on all smoke types being handled as blue.
-        if (new_cloud == CLOUD_BLUE_SMOKE)
-            clouds_list.push_back("coloured smoke");
-        else
-            clouds_list.push_back(cloud_type_name((cloud_type) new_cloud));
-
-        old_cloud = new_cloud;
-    }
-
-    return comma_separated_line(clouds_list.begin(), clouds_list.end(),
-                                " or ", ", ");
-}
-
-
-// Assumes beem.range is already set -cao
-spret_type cast_evaporate(int pow, bolt& beem, int pot_idx, bool fail)
-{
-    ASSERT(you.inv[pot_idx].base_type == OBJ_POTIONS);
-    item_def& potion = you.inv[pot_idx];
-
-    beem.name        = "potion";
-    beem.colour      = potion.colour;
-    beem.glyph       = dchar_glyph(DCHAR_FIRED_FLASK);
-    beem.beam_source = MHITYOU;
-    beem.thrower     = KILL_YOU_MISSILE;
-    beem.is_beam     = false;
-    beem.aux_source.clear();
-
-    beem.auto_hit   = true;
-    beem.damage     = dice_def(1, 0);  // no damage, just producing clouds
-    beem.ench_power = pow;               // used for duration only?
-    beem.is_explosion = true;
-
-    beem.flavour    = BEAM_POTION_MEPHITIC;
-    beam_type tracer_flavour = BEAM_MMISSILE;
-
-    switch (potion.sub_type)
-    {
-    case POT_STRONG_POISON:
-        beem.ench_power *= 2;
-        // deliberate fall-through
-    case POT_POISON:
-        beem.flavour   = BEAM_POTION_POISON;
-        tracer_flavour = BEAM_POISON;
-        break;
-
-    case POT_DEGENERATION:
-        beem.effect_known = false;
-        beem.flavour   = (coinflip() ? BEAM_POTION_POISON : BEAM_POTION_MIASMA);
-        tracer_flavour = BEAM_MIASMA;
-        beem.ench_power *= 2;
-        break;
-
-    case POT_DECAY:
-        beem.flavour   = BEAM_POTION_MIASMA;
-        tracer_flavour = BEAM_MIASMA;
-        beem.ench_power *= 2;
-        break;
-
-    case POT_PARALYSIS:
-        beem.ench_power *= 2;
-        // fall through
-    case POT_CONFUSION:
-    case POT_SLOWING:
-        tracer_flavour = beem.flavour = BEAM_POTION_MEPHITIC;
-        break;
-
-    case POT_WATER:
-    case POT_PORRIDGE:
-        tracer_flavour = beem.flavour = BEAM_POTION_STEAM;
-        break;
-
-    case POT_BLOOD:
-    case POT_BLOOD_COAGULATED:
-        if (one_chance_in(3))
-            break; // stinking cloud
-        // deliberate fall through
-    case POT_BERSERK_RAGE:
-        beem.effect_known = false;
-        beem.flavour = (coinflip() ? BEAM_POTION_FIRE : BEAM_POTION_STEAM);
-        if (potion.sub_type == POT_BERSERK_RAGE)
-            tracer_flavour = BEAM_FIRE;
-        else
-            tracer_flavour = BEAM_RANDOM;
-        break;
-
-    case POT_MUTATION:
-        // Maybe we'll get a mutagenic cloud.
-        if (coinflip())
-        {
-            beem.effect_known = true;
-            tracer_flavour = beem.flavour = BEAM_POTION_MUTAGENIC;
-            break;
-        }
-        // if not, deliberate fall through for something random
-
-    case POT_GAIN_STRENGTH:
-    case POT_GAIN_DEXTERITY:
-    case POT_GAIN_INTELLIGENCE:
-    case POT_EXPERIENCE:
-    case POT_MAGIC:
-        beem.effect_known = false;
-        switch (random2(5))
-        {
-        case 0:   beem.flavour = BEAM_POTION_FIRE;   break;
-        case 1:   beem.flavour = BEAM_POTION_COLD;   break;
-        case 2:   beem.flavour = BEAM_POTION_POISON; break;
-        case 3:   beem.flavour = BEAM_POTION_MIASMA; break;
-        default:  beem.flavour = BEAM_POTION_RANDOM; break;
-        }
-        tracer_flavour = BEAM_RANDOM;
-        break;
-
-    default:
-        beem.effect_known = false;
-        switch (random2(12))
-        {
-        case 0:   beem.flavour = BEAM_POTION_FIRE;            break;
-        case 1:   beem.flavour = BEAM_POTION_MEPHITIC;        break;
-        case 2:   beem.flavour = BEAM_POTION_COLD;            break;
-        case 3:   beem.flavour = BEAM_POTION_POISON;          break;
-        case 4:   beem.flavour = BEAM_POTION_RANDOM;          break;
-        case 5:   beem.flavour = BEAM_POTION_BLUE_SMOKE;      break;
-        case 6:   beem.flavour = BEAM_POTION_BLACK_SMOKE;     break;
-        default:  beem.flavour = BEAM_POTION_STEAM;           break;
-        }
-        tracer_flavour = BEAM_RANDOM;
-        break;
-    }
-
-    // Fire tracer. FIXME: use player_tracer() here!
-    beem.source         = you.pos();
-    beem.can_see_invis  = you.can_see_invisible();
-    beem.smart_monster  = true;
-    beem.attitude       = ATT_FRIENDLY;
-    beem.beam_cancelled = false;
-    beem.is_tracer      = true;
-    beem.friend_info.reset();
-
-    beam_type real_flavour = beem.flavour;
-    beem.flavour           = tracer_flavour;
-    beem.fire();
-
-    if (beem.beam_cancelled)
-    {
-        // We don't want to fire through friendlies or at ourselves.
-        canned_msg(MSG_OK);
-        return SPRET_ABORT;
-    }
-
-    fail_check();
-    // Really fire.
-    beem.flavour = real_flavour;
-    beem.is_tracer = false;
-    beem.fire();
-
-    // Use up a potion.
-    if (is_blood_potion(potion))
-        remove_oldest_blood_potion(potion);
-
-    dec_inv_item_quantity(pot_idx, 1);
-
-    return SPRET_SUCCESS;
 }
 
 int holy_flames(monster* caster, actor* defender)
